Move multipart helper into shared utils module
