+++ nooneissafe/notification_providers/discord.py
@@ -1,7 +1,7 @@
 import json
 import logging
 
-from .multipart import post_multipart
+from ..utils import post_multipart
 
 
 logger = logging.getLogger(__name__)
